requirements.txt
# Nike Rocket Follower API - Requirements
# =========================================

# FastAPI framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# HTTP requests
requests==2.31.0
aiohttp>=3.10.11

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment
python-dotenv==1.0.0

# Math/Stats (for portfolio tracking)
numpy==1.26.2

# Encryption (NEW!)
cryptography==41.0.7

asyncpg==0.29.0

# Hyperliquid SDK
hyperliquid-python-sdk>=0.8.1
eth-account>=0.10.0

slowapi==0.1.9

# Email
resend>=0.5.0
